requirements.txt
# BAMSec Filing Bot - preferreds / CEFs SEC filings + Telegram
requests>=2.31.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
